+++ web/src/navigation/app_routes.ts
@@ -4,37 +4,17 @@ import Login from "@/pages/Login";
 import NotFound from "@/pages/NotFound";
 import Unauthorized from "@/pages/Unauthorized";
 import Dashboard from "@/pages/dashboard";
-import MedicalTests from "@/pages/medicalTests";
-import {
-  AddMedicalTest,
-  type AddMedicalTestProps,
-} from "@/pages/medicalTests/AddMedicalTest";
-import {
-  EditMedicalTest,
-  type EditMedicalTestProps,
-} from "@/pages/medicalTests/EditMedicalTest";
-import {
-  ViewMedicalTest,
-  type ViewMedicalTestProps,
-} from "@/pages/medicalTests/ViewMedicalTest";
 import Profile from "@/pages/profile";
-import Users from "@/pages/users";
-import AddUser from "@/pages/users/AddUser";
-import EditUser from "@/pages/users/EditUser";
-import ViewUser from "@/pages/users/ViewUser";
 import {
-  EditIcon,
-  EyeIcon,
   LayoutDashboardIcon,
   LogInIcon,
-  MicroscopeIcon,
-  PlusSquareIcon,
   ShieldBanIcon,
   TriangleAlertIcon,
   UserCogIcon,
-  UsersIcon,
   type LucideProps,
 } from "lucide-react";
+import { MedicalTestsRoute } from "./medical-tests-routes";
+import { UsersRoute } from "./users-routes";
 
 type RouteType = "auth" | "public" | "protected";
 
@@ -44,7 +24,7 @@ export interface AppRoute<P = {}> {
   icon: React.ForwardRefExoticComponent<
     Omit<LucideProps, "ref"> & React.RefAttributes<SVGSVGElement>
   >;
-  route: React.ComponentType<P>;
+  route?: React.ComponentType<P>;
   layout?: React.ComponentType;
   type: RouteType;
   redirectTo?: string;
@@ -55,87 +35,28 @@ export interface AppRoute<P = {}> {
   hideSidebarToggle?: boolean;
 }
 
-export const AddMedicalTestRoute: AppRoute<AddMedicalTestProps> = {
-  path: "/medical-tests/new",
-  label: "Add a medical test",
-  icon: PlusSquareIcon,
-  route: AddMedicalTest,
-  type: "protected",
-  requiredRoles: ["admin", "super_admin"],
-  requiredPermissions: ["medical_tests.create"],
-};
+function buildBase(props: AppRoute): AppRoute {
+  return { ...props };
+}
 
-export const ViewMedicalTestRoute: AppRoute<ViewMedicalTestProps> = {
-  path: "/medical-tests/:id",
-  label: "Medical Test Details",
-  icon: EyeIcon,
-  route: ViewMedicalTest,
-  type: "protected",
-  requiredRoles: ["user", "admin", "super_admin"],
-  requiredPermissions: ["medical_tests.read"],
-};
+export function createRouteWithChildren(
+  params: Omit<AppRoute, "layout" | "children" | "route"> & {
+    layout: React.ComponentType;
+    children: AppRoute<any>[];
+  }
+): AppRoute {
+  return {
+    ...params,
+  };
+}
 
-export const UpdateMedicalTestRoute: AppRoute<EditMedicalTestProps> = {
-  path: "/medical-tests/edit/:id",
-  label: "Edit Medical Test Details",
-  icon: EditIcon,
-  route: EditMedicalTest,
-  type: "protected",
-  requiredRoles: ["admin", "super_admin"],
-  requiredPermissions: ["medical_tests.update"],
-};
-
-export const MedicalTestsRoute: AppRoute = {
-  path: "/medical-tests",
-  label: "Medical Tests",
-  icon: MicroscopeIcon,
-  route: MedicalTests,
-  type: "protected",
-  requiredRoles: ["user", "admin", "super_admin"],
-  requiredPermissions: ["medical_tests.read"],
-  children: [AddMedicalTestRoute, ViewMedicalTestRoute, UpdateMedicalTestRoute],
-};
-
-export const AddUserRoute: AppRoute = {
-  path: "/users/new",
-  label: "Add a user",
-  icon: PlusSquareIcon,
-  route: AddUser,
-  type: "protected",
-  requiredRoles: ["super_admin"],
-  requiredPermissions: ["users.create"],
-};
-
-export const ViewUserRoute: AppRoute = {
-  path: "/users/:id",
-  label: "User's Details",
-  icon: EyeIcon,
-  route: ViewUser,
-  type: "protected",
-  requiredRoles: ["admin", "super_admin"],
-  requiredPermissions: ["users.read"],
-};
-
-export const UpdateUserRoute: AppRoute = {
-  path: "/users/edit/:id",
-  label: "Edit User's Details",
-  icon: EditIcon,
-  route: EditUser,
-  type: "protected",
-  requiredRoles: ["super_admin"],
-  requiredPermissions: ["users.update"],
-};
-
-export const UsersRoute: AppRoute = {
-  path: "/users",
-  label: "Users",
-  icon: UsersIcon,
-  route: Users,
-  type: "protected",
-  requiredRoles: ["admin", "super_admin"],
-  requiredPermissions: ["users.read"],
-  children: [AddUserRoute, ViewUserRoute, UpdateUserRoute],
-};
+export function createLeafRoute<P = {}>(
+  params: Omit<AppRoute<P>, "layout" | "children">
+): AppRoute<P> {
+  return {
+    ...params,
+  };
+}
 
 export const LoginRoute: AppRoute = {
   path: "/",
