+++ web/src/pages/dashboard/index.tsx
@@ -12,8 +12,9 @@ import {
 } from "@/components/ui/item";
 import { Spinner } from "@/components/ui/spinner";
 import { useAuth } from "@/context/AuthContext";
-import { ProfileRoute, UsersRoute } from "@/navigation/app_routes";
+import { ProfileRoute } from "@/navigation/app_routes";
 import { canAccessRoute } from "@/navigation/guards";
+import { UsersRoute } from "@/navigation/users-routes";
 import { getUserLastConnectionById } from "@/services/SupabaseService";
 import { UserCogIcon } from "lucide-react";
 import { useEffect, useState } from "react";
