+++ web/src/pages/users/index.tsx
@@ -1,12 +1,7 @@
-import ListTitle from "@/components/shared/ListTitle";
-import { AdminsData } from "@/shared/entity-data";
+import ParentRouteLayoutTemplate from "@/components/routing/ParentRouteLayoutTemplate";
 
-const Users = ({}) => {
-  return (
-    <div className="p-5 space-y-5">
-      <ListTitle title={AdminsData.title} description={AdminsData.title} />
-    </div>
-  );
+const UsersLayout = ({}) => {
+  return <ParentRouteLayoutTemplate />;
 };
 
-export default Users;
+export default UsersLayout;
